Uploaded calculator.py
This file contains functions to calculate the chance of rolling above, below, or equal to a given amount on a roll of some number X of Y-sided dice. Optionally, these functions can also consider a flat "bonus" to this roll, Z, but for the purposes of these functions, this is just an offset to the target amount.

Credit to user G Cab on Stack Overflow for providing the mathematical formula for calculating the number of outcomes where the die roll is equal to a target number. Link: https://math.stackexchange.com/questions/992125/if-i-roll-three-dice-at-the-same-time-how-many-ways-the-sides-can-sum-up-to-13/1680420#1680420